refactor: moved the filename that get TFunction
